requirements.txt
streamlit
google-genai
bs4